Factor out `Name` into its own (re-exported) module.
This is in preparation for hiding the `unName` and `unNameSuffix`
functions. There are very few reasons to deconstruct a `Name` or
`NameSuffix`; we may be able to eliminate all of them.
