+++ src/Language/GraphQL/Draft/Parser.hs-boot
@@ -8,9 +8,10 @@ where
 
 import Data.Text (Text)
 import {-# SOURCE #-} Language.GraphQL.Draft.Syntax qualified as AST
+import Language.GraphQL.Draft.Syntax.Name (Name)
 import Prelude (Either)
 
 -------------------------------------------------------------------------------
 
-parseExecutableDoc :: Text -> Either Text (AST.ExecutableDocument AST.Name)
+parseExecutableDoc :: Text -> Either Text (AST.ExecutableDocument Name)
 parseSchemaDocument :: Text -> Either Text AST.SchemaDocument
