+++ src/Language/GraphQL/Draft/Syntax/Name.hs-boot
@@ -1,0 +1,8 @@
+module Language.GraphQL.Draft.Syntax.Name (Name) where
+
+import Data.Kind (Type)
+
+-------------------------------------------------------------------------------
+
+type Name :: Type
+data Name
